Harden v0.2.3 performance and warning gates
