+++ Cargo.toml
@@ -9,7 +9,7 @@ resolver = "3"
 [workspace.package]
 edition = "2024"
 license = "MIT"
-version = "0.2.2"
+version = "0.2.3"
 rust-version = "1.95"
 
 [workspace.dependencies]
