+++ crates/netdiag-core/Cargo.toml
@@ -31,3 +31,7 @@ etherparse.workspace = true
 float-cmp.workspace = true
 insta.workspace = true
 tempfile.workspace = true
+
+[[bench]]
+name = "perf_budget"
+harness = false
